Update merge conflict or rename file
